lista para pruebas

+++ app/Image.php
@@ -18,12 +18,15 @@ class Image extends Model
     
     public static function deleteImage($image_id){
         
-        $image = Image::find($image_id);
+         $image = Image::find($image_id);
+        if(sizeof($image->article->images) > 1){
         unlink(public_path()."\images\articles\\".$image->image_url);
-         /* unlink("/home/eselenas/public_html/images/articles/".$image->image_url); asi se elimina en hosting*/
         $image->delete();
-        
+        /* unlink("/home2/dsistema/public_html/images/articles/".$image->image_url);*/
         return Flash::success("Imagen eliminada");
+            }else{
+             return Flash::success("El artículo no puede quedar sin imagen");
+        }
     }
     
     public static function uploadImage($article, $request){
